Move AdvAcrobot to its own file and clear up naming/concepts
