+++ gym_rarl/envs/adv_acrobot.py
@@ -1,49 +1,17 @@
-import abc
-
-import gym
 from gym.envs.classic_control.acrobot import *
 
-
-class BaseAdversarialEnvWrapper(abc.ABC, gym.core.Env):
-    """
-    Base environment wrapping a Gym task where both a main agent and adversary act in each step
-    """
-
-    def __init__(self):
-        super().__init__()
-        self.adv_agent = None
-        self.main_agent = None
-
-    def link_agents(self, main_agent, adv_agent):
-        self.main_agent = main_agent
-        self.adv_agent = adv_agent
-
-    def is_linked(self):
-        """Returns whether this environment linked with both the main and adversarial agent"""
-        return self.adv_agent is not None and self.main_agent is not None
-
-    @abc.abstractmethod
-    def get_ob(self):
-        """
-        Return the latest observation s_t
-        """
-        pass
-
-    @abc.abstractmethod
-    def step_two_actions(self, main_action, adv_action):
-        """
-        Step the environment with an action from both the main agent and the adversarial agent.
-        """
-        pass
+from gym_rarl.envs.adv_env import BaseAdversarialEnvWrapper
 
 
 class AdversarialAcrobotEnvWrapper(BaseAdversarialEnvWrapper, AcrobotEnv):
-    """Base env for Acrobot task which bridges between main and adversarial agent"""
+    """
+    Base env for Acrobot task which allows main/adv to act in each step
+    """
 
     def get_ob(self):
         return self._get_ob()
 
-    def step_two_actions(self, main_action, adv_action):
+    def step_two_actors(self, main_action, adv_action):
         """
         Copied from AcrobotEnv (gym==0.17.3)
         """
